Soft delete with paranoia

+++ app/controllers/backend/products_controller.rb
@@ -3,7 +3,7 @@ class Backend::ProductsController < Backend::BaseController
   before_action :load_product, except: %i(index new create import)
 
   def index
-    @q = Product.ransack params[:q]
+    @q = Product.with_deleted.ransack params[:q]
     @products = @q.result.includes(:category).newest
                   .paginate page: params[:page],
                    per_page: Settings.admin_product_perpage
@@ -37,10 +37,27 @@ class Backend::ProductsController < Backend::BaseController
   end
 
   def destroy
-    if @product.destroy
+    if @product.really_destroy!
       flash[:success] = t ".delete_success"
     else
       flash[:danger] = t ".delete_fail"
+    end
+    redirect_to backend_products_path
+  end
+
+  def softdelete
+    if @product.deleted?
+      if @product.restore recursive: true
+        flash[:success] = t ".restore_success"
+      else
+        flash[:danger] = t ".restore_fail"
+      end
+    else
+      if @product.destroy
+        flash[:success] = t ".soft_del_success"
+      else
+        flash[:danger] = t ".soft_del_fail"
+      end
     end
     redirect_to backend_products_path
   end
@@ -72,7 +89,7 @@ class Backend::ProductsController < Backend::BaseController
   end
 
   def load_product
-    @product = Product.find params[:id]
+    @product = Product.with_deleted.find params[:id]
   end
 
   def search_key_cat
